changes to tests file,created item.js, and  updated file paths in app.js
