+++ models/index.js
@@ -1,18 +1,13 @@
 const {Sequelize, DataTypes, Model} = require('sequelize')
 const {sequelize} = require('../db')
+const {Warehouse} = require('./warehouse');
+const {Item} = require('./item');
 
 
-class Item extends Model {} //change to Inventory models
 
-Item.init({                 //change to Inventory models
-    title: DataTypes.STRING,
-    price: DataTypes.FLOAT(2),
-    description: DataTypes.TEXT,
-    category: DataTypes.STRING,
-    image: DataTypes.STRING,
-}, {
-    sequelize,
-    timestamps: false,
-});
 
-module.exports = {Item}; //change to export Inventory models
+
+
+
+
+module.exports = {Item, Warehouse}; //change to export Inventory models
