+++ models/index.test.js
@@ -1,5 +1,7 @@
 const {Item} = require('./index'); //change to Inventory models
 const {sequelize} = require('../db');
+const {Warehouse} = require('./warehouse'); //change to Inventory models
+
 
 describe('Item Model', () => {		//change test to Inventory models
 	beforeAll(async () => {
@@ -16,5 +18,20 @@ describe('Item Model', () => {		//change test to Inventory models
 		expect(testItem.image).toBe('url.png')
 	})
 
+
+})
+
+describe('Warehouse Model', () => {		//change test to Inventory models
+	beforeAll(async () => {
+		await sequelize.sync({force: true})
+	});
+
+	test('can create a Warehouse', async() => {
+		const testWarehouse = await Warehouse.create({name : 'Macys', location : 'Irving'})
+		expect(testWarehouse.name).toBe('Macys')
+		expect(testWarehouse.location).toBe('Irving')
+		expect(testWarehouse.id).toBe(1)
+		console.log(testWarehouse.id)
+	})
 
 })
